That's how I intended to do GPGPU version of filters, before I ran into the issue with double precision not being supported.

+++ src/hw-1/clfilters.h
@@ -4,6 +4,7 @@
 #define CL_TARGET_OPENCL_VERSION 120
 
 #include <gtk/gtk.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <CL/cl.h>
@@ -13,11 +14,11 @@ void host_program(guint8 *,
                 size_t ,
                 int ,
                 int ,
-                float *, 
+                double *, 
                 int ,
                 int ,
                 int ,
-                float ,
-                float);
+                double ,
+                double);
 
 #endif /* __CLFILTERS__ */
